apps/gateway/requirements.txt
# Gateway Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
authlib>=1.0.0
httpx>=0.25.0
python-jose>=3.3.0
passlib>=1.7.0
bcrypt>=4.0.0
python-multipart>=0.0.6
sse-starlette>=1.6.0
